adaptive_tutor/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adaptive-tutor"
version = "0.1.0"
description = "Adaptive tutoring backend with LangGraph, Gemini, and FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "google-genai>=1.10.0",
  "langchain>=0.3.0",
  "langgraph>=0.2.30",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "python-dotenv>=1.0.1",
  "sqlalchemy>=2.0.30",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
